requirements.txt
Pillow==7.2.0
gpxpy==1.4.2
